Fix query endpoint to return more than just doc number. Include doc number in mapped response.
